Factor out ReadSnapshot from UniqueStoreDictionary.
