+++ vespalib/src/vespa/vespalib/datastore/unique_store_dictionary.h
@@ -8,6 +8,7 @@
 namespace vespalib::datastore {
 
 class EntryComparatorWrapper;
+class IUniqueStoreDictionaryReadSnapshot;
 
 class NoHashDictionary;
 
@@ -56,19 +57,7 @@ protected:
     using BTreeDictionaryType = BTreeDictionaryT;
     using DataType = typename BTreeDictionaryType::DataType;
     using FrozenView = typename BTreeDictionaryType::FrozenView;
-    using ReadSnapshot = typename ParentT::ReadSnapshot;
     using generation_t = typename ParentT::generation_t;
-
-    class ReadSnapshotImpl : public ReadSnapshot {
-    private:
-        FrozenView _frozen_view;
-
-    public:
-        ReadSnapshotImpl(FrozenView frozen_view);
-        size_t count(const EntryComparator& comp) const override;
-        size_t count_in_range(const EntryComparator& low, const EntryComparator& high) const override;
-        void foreach_key(std::function<void(EntryRef)> callback) const override;
-    };
 
 public:
     using UniqueStoreBTreeDictionaryBase<BTreeDictionaryT>::has_btree_dictionary;
@@ -87,7 +76,7 @@ public:
     void build(vespalib::ConstArrayRef<EntryRef>, vespalib::ConstArrayRef<uint32_t> ref_counts, std::function<void(EntryRef)> hold) override;
     void build(vespalib::ConstArrayRef<EntryRef> refs) override;
     void build_with_payload(vespalib::ConstArrayRef<EntryRef>, vespalib::ConstArrayRef<uint32_t> payloads) override;
-    std::unique_ptr<ReadSnapshot> get_read_snapshot() const override;
+    std::unique_ptr<IUniqueStoreDictionaryReadSnapshot> get_read_snapshot() const override;
     bool get_has_hash_dictionary() const override;
 };
